+++ modules/edhoc/src/th.c
@@ -37,7 +37,7 @@ th2_input_encode(uint8_t *hash_msg1, uint32_t hash_msg1_len, uint8_t *g_y,
 		 uint32_t g_y_len, struct c_x *c_r, uint8_t *th2_input,
 		 uint16_t *th2_input_len)
 {
-	size_t payload_len_out;
+	uint32_t payload_len_out;
 	struct th2 th2;
 
 	/*Encode hash_msg1*/
@@ -96,7 +96,7 @@ static inline enum edhoc_error th3_input_encode(uint8_t *th2, uint8_t th2_len,
 	th3._th3_CIPHERTEXT_2.value = ciphertext_2;
 	th3._th3_CIPHERTEXT_2.len = ciphertext_2_len;
 
-	size_t payload_len_out;
+	uint32_t payload_len_out;
 	TRY_EXPECT(cbor_encode_th3(th3_input, *th3_input_len, &th3,
 				   &payload_len_out),
 		   true);
@@ -132,7 +132,7 @@ static inline enum edhoc_error th4_input_encode(uint8_t *th3, uint8_t th3_len,
 	th4._th4_CIPHERTEXT_3.value = ciphertext_3;
 	th4._th4_CIPHERTEXT_3.len = ciphertext_3_len;
 
-	size_t payload_len_out;
+	uint32_t payload_len_out;
 	TRY_EXPECT(cbor_encode_th4(th4_input, *th4_input_len, &th4,
 				   &payload_len_out),
 		   true);
